% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\docType{class}
\name{abc}
\alias{abc}
\alias{abcd.ppomp-class}
\alias{abc,panelPomp-method}
\title{PABC: Panel Approximate Bayesian Computation}
\usage{
\S4method{abc}{panelPomp}(object, Nabc = 1, start, proposal, probes, scale,
  epsilon, verbose = getOption("verbose"), pproposal = NULL,
  pprobes = NULL, pscale = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{panelPomp} or inheriting class 
\code{panelPomp}.}

\item{Nabc}{
    The number of ABC iterations to perform.
  }

\item{start}{position in original \code{times(pomp)} at which to start.}

\item{proposal}{
    optional function that draws from the proposal distribution.
    Currently, the proposal distribution must be symmetric for proper inference:
    it is the user's responsibility to ensure that it is.
    Several functions that construct appropriate proposal function are provided:
    see \link{MCMC proposal functions} for more information.
  }

\item{probes}{
    List of probes (AKA summary statistics).
    See \code{\link{probe}} for details.
  }

\item{scale}{
    named numeric vector of scales.
  }

\item{epsilon}{
    ABC tolerance.
  }

\item{verbose}{
    logical; if TRUE, print progress reports.
  }

\item{pproposal}{\inheritArg{proposal}{abc}.}

\item{pprobes}{\inheritArg{probe}{abc}.}

\item{pscale}{\inheritArg{scale}{abc}.}

\item{...}{....}
}
\description{
Tools for applying Approximate Bayesian Computation algorithms
to panel data.

\inheritDes{abc}.
}
\references{
\tavare1997

\sisson2007

\sisson2009
}
\seealso{
\pkg{pomp}'s abc at \link[=abc,pomp-method]{abc}
}
