% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelPomp_methods.R
\docType{methods}
\name{coef<-,panelPomp-method}
\alias{coef<-,panelPomp-method}
\title{'coef<-' method for panelPomp signature.}
\usage{
\S4method{coef}{panelPomp}(object, ...) <- value
}
\arguments{
\item{object}{A \code{panelPomp} object.}

\item{...}{....}

\item{value}{value.}
}
\description{
'coef<-' method for panelPomp signature.
}
\author{
Carles Breto and Aaron A. King.
}
