% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacts.R
\name{contacts}
\alias{contacts}
\title{Contacts model}
\usage{
contacts(
  params = c(mu_X = 1.75, sigma_X = 2.67, mu_D = 3.81, sigma_D = 4.42, mu_R = 0.04,
    sigma_R = 0, alpha = 0.9)
)
}
\arguments{
\item{params}{parameter vector.}
}
\description{
A panel model for dynamic variation in sexual contacts, with
data from Vittinghof et al (1999). The model was developed by
Romero-Severson et al (2015) and discussed by Breto et al (2019).
}
\examples{
contacts()
}
\references{
\breto2018

\vittinghoff1999

\romeroseverson2015
}
\author{
Edward L. Ionides
}
