% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mif2.R
\name{.modifyOther}
\alias{.modifyOther}
\title{Internal function for modifying pparamArray in Mif2}
\usage{
.modifyOther(x, spnames, indices, unit)
}
\arguments{
\item{x}{pparamArray in mif2 internal}

\item{spnames}{names of parameters to update}

\item{indices}{indices that are the output of a mfi2 call to pomp}

\item{unit}{unit in the unit loop}
}
\description{
Internal function for modifying pparamArray in Mif2
}
