% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mif2.R
\name{.modifySelf}
\alias{.modifySelf}
\title{Internal function for modifying pparamArray in Mif2}
\usage{
.modifySelf(x, spnames, M, unit)
}
\arguments{
\item{x}{pparamArray in mif2 internal}

\item{spnames}{names of parameters to update}

\item{M}{Matrix of replacement values}

\item{unit}{unit in the unit loop}
}
\description{
Internal function for modifying pparamArray in Mif2
}
