% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-exported.R
\name{get_col}
\alias{get_col}
\title{Get column.}
\usage{
get_col(matrix, rows, col)
}
\arguments{
\item{matrix}{matrix.}

\item{rows}{numeric; rows to subset; like with `[`, this argument can be left empty to designate all rows.}

\item{col}{numeric; single column to subset.}
}
\description{
Subset matrix dropping dimension but without dropping rownames (which is R's default).
}
