% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-exported.R
\name{get_row}
\alias{get_row}
\title{Get row.}
\usage{
get_row(matrix, row, cols)
}
\arguments{
\item{matrix}{matrix.}

\item{row}{numeric; single row to subset.}

\item{cols}{numeric; columns to subset; like with `[`, this argument can be left empty to designate all columns.}
}
\description{
Subset matrix dropping dimension but without dropping colnames (which is R's default).
}
