% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_gompertz.R
\name{kalman_gompertz}
\alias{kalman_gompertz}
\title{Gompertz Kalman filter.}
\usage{
kalman_gompertz(Y, X0, r, K, sigma, tau, from_est = FALSE, po = NULL)
}
\arguments{
\item{Y}{data.}

\item{X0}{initial value for state variable X.}

\item{r}{r parameter.}

\item{K}{K parameter.}

\item{sigma}{sigma parameter.}

\item{tau}{tau parameter.}

\item{from_est}{logical; whether to partrans with from_est before the final calculation (e.g., if passing to optim as opposed to evaluating directly).}

\item{po}{pomp; Gompertz object to use with partrans (if transform is TRUE).}
}
\description{
Gompertz Kalman filter.
}
