% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_gompertz.R
\name{kalman_gompertz_u}
\alias{kalman_gompertz_u}
\title{Kalman filter for a Gompertz pomp object.}
\usage{
kalman_gompertz_u(x, po, params, xnm = NULL, from_est = FALSE)
}
\arguments{
\item{x}{parameter values over which to optimize.}

\item{po}{pomp; Gompertz object to use with partrans (if transform is TRUE).}

\item{params}{numeric vector with values for parameter not specified in x.}

\item{xnm}{Including the names of x in xnm allows reassigning them if dropped by optim (as happens with the Brent method for in univariate optimization).}

\item{from_est}{logical; whether to partrans with from_est before the final calculation (e.g., if passing to optim as opposed to evaluating directly).}
}
\description{
Kalman filter for a Gompertz pomp object.
}
