% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_gompertz.R
\name{kalman_panel}
\alias{kalman_panel}
\title{Evaluates the Kalman filter loglikelihood of a panel Gompertz.}
\usage{
kalman_panel(
  sh,
  obj,
  pParams,
  unit_fn,
  shnm = NULL,
  allout = FALSE,
  from_est = FALSE
)
}
\arguments{
\item{sh}{shared parameters values.}

\item{obj}{panelPomp object.}

\item{pParams}{list of shared and specific parameters.}

\item{unit_fn}{function to evaluate the unit loglikelihood.}

\item{shnm}{shared parameter names; needed because, if using optim's Brent (as recommended by optim when there is only one shared parameter), optim strips these off the sh argument.}

\item{allout}{logical; whether to return all the optimization output (TRUE) or only the optimized values (FALSE, as needed when passed to \code{optim}).}

\item{from_est}{logical; whether to partrans with from_est before the final calculation (e.g., if passing to optim as opposed to evaluating directly).}
}
\description{
Evaluates the Kalman filter loglikelihood of a panel Gompertz.
}
