% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelPomp-methods.R
\docType{methods}
\name{length,panelPomp-method}
\alias{length,panelPomp-method}
\title{Count the number of units in the \code{unitobjects} slot of \code{panelPomp} objects.}
\usage{
\S4method{length}{panelPomp}(x)
}
\arguments{
\item{x}{A \code{panelPomp} object.}
}
\description{
Help me!
}
\details{
S4 method.
}
