% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfilterd-ppomp-methods.R
\docType{methods}
\name{logLik,pfilterd.ppomp-method}
\alias{logLik,pfilterd.ppomp-method}
\title{Extract loglikelihood from \code{pfilterd.ppomp} object}
\usage{
\S4method{logLik}{pfilterd.ppomp}(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{Additional arguments}
}
\description{
Now then, now then, what's all this?
}
\details{
S4 method
}
