% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-polio.R
\docType{data}
\name{longlat}
\alias{longlat}
\title{Longitudes and latitudes for the polio data}
\format{An object of class \code{data.frame} with 49 rows and 4 columns.}
\usage{
longlat
}
\description{
Longitudes and latitudes for the polio data
}
\keyword{datasets}

