% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mif2d-ppomp-methods.R
\docType{methods}
\name{mif2,mif2d.ppomp-method}
\alias{mif2,mif2d.ppomp-method}
\title{Apply the \code{mif2} algorithm to \code{mif2d.ppomp} objects.}
\usage{
\S4method{mif2}{mif2d.ppomp}(object, Nmif, shared.start, specific.start, Np,
  prw.sd, transform, cooling.type, cooling.fraction.50, ptol, ...)
}
\arguments{
\item{object}{
    An object of class \code{pomp}.
  }

\item{Nmif}{
    The number of filtering iterations to perform.
  }

\item{shared.start}{shared.arg.}

\item{specific.start}{specific.arg.}

\item{Np}{
    the number of particles to use in filtering.
    This may be specified as a single positive integer, in which case the same number of particles will be used at each timestep.
    Alternatively, if one wishes the number of particles to vary across timestep, one may specify \code{Np} either as a vector of positive integers (of length \code{length(time(object))}) or as a function taking a positive integer argument.
    In the latter case, \code{Np(n)} must be a single positive integer, representing the number of particles to be used at the \code{n}-th timestep:
    \code{Np(1)} is the number of particles to use going from \code{timezero(object)} to \code{time(object)[1]},
    \code{Np(2)}, from \code{time(object)[1]} to \code{time(object)[2]},
    and so on.
    \strong{Note that this behavior differs from that of \code{\link{mif}}!}
  }

\item{prw.sd}{An unevaluated expression of the form \code{quote(rw.sd())} to be used for all panel units. If a \code{list} of such expressions of the same length as the \code{object} argument is provided, each list element will be used for the corresponding panel unit.}

\item{transform}{transform (seems to cause an error if documentation inherited from 'pomp' package)}

\item{cooling.type}{
    specifications for the cooling schedule, i.e., the manner in which the intensity of the parameter perturbations is reduced with successive filtering iterations.
    \code{cooling.type} specifies the nature of the cooling schedule.
    See below (under \dQuote{Specifying the perturbations}) for more detail.
  }

\item{cooling.fraction.50}{cooling.fraction.50 (seems to cause an error if documentation inherited from 'pomp' package)}

\item{ptol}{Unit-specific tolerances can be passed as a named numeric vector with names matching \code{names(unitobjects(object))}.}

\item{...}{...}
}
\description{
S4 method
}
\details{
S4 method
}

