% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelPomp-methods.R
\docType{methods}
\name{mif2,panelPomp-method}
\alias{mif2,panelPomp-method}
\title{Apply the \code{mif2} algorithm to a \code{panelPomp} object.}
\usage{
\S4method{mif2}{panelPomp}(object, shared.start = numeric(0),
  specific.start = array(data = numeric(0), dim = c(0, 0)), Np, Nmif,
  cooling.type, cooling.fraction.50, transform, prw.sd,
  verbose = getOption("verbose"), ...)
}
\arguments{
\item{object}{A \code{panelPomp} object.}

\item{shared.start}{shared.arg.}

\item{specific.start}{specific.arg.}

\item{Np}{
    the number of particles to use in filtering.
    This may be specified as a single positive integer, in which case the same number of particles will be used at each timestep.
    Alternatively, if one wishes the number of particles to vary across timestep, one may specify \code{Np} either as a vector of positive integers (of length \code{length(time(object))}) or as a function taking a positive integer argument.
    In the latter case, \code{Np(n)} must be a single positive integer, representing the number of particles to be used at the \code{n}-th timestep:
    \code{Np(1)} is the number of particles to use going from \code{timezero(object)} to \code{time(object)[1]},
    \code{Np(2)}, from \code{time(object)[1]} to \code{time(object)[2]},
    and so on.
    \strong{Note that this behavior differs from that of \code{\link{mif}}!}
  }

\item{Nmif}{
    The number of filtering iterations to perform.
  }

\item{cooling.type}{
    specifications for the cooling schedule, i.e., the manner in which the intensity of the parameter perturbations is reduced with successive filtering iterations.
    \code{cooling.type} specifies the nature of the cooling schedule.
    See below (under \dQuote{Specifying the perturbations}) for more detail.
  }

\item{cooling.fraction.50}{cooling.fraction.50 (seems to cause an error if documentation inherited from 'pomp' package)}

\item{transform}{transform (seems to cause an error if documentation inherited from 'pomp' package)}

\item{prw.sd}{An unevaluated expression of the form \code{quote(rw.sd())} to be used for all panel units. If a \code{list} of such expressions of the same length as the \code{object} argument is provided, each list element will be used for the corresponding panel unit.}

\item{verbose}{
    logical; if TRUE, print progress reports.
  }

\item{...}{
    additional arguments that override the defaults.
  }
}
\description{
S4 method.
}
\details{
S4 method.
}

