% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mif2d-ppomp-class.R
\docType{class}
\name{mif2d.ppomp-class}
\alias{mif2d.ppomp-class}
\title{An S4 class to represent panel POMP models that have been analyzed using \code{mif2}}
\description{
An S4 class to represent panel POMP models that have been analyzed using \code{mif2}
}
\section{Slots}{

\describe{
\item{\code{Nmif}}{Nmif}

\item{\code{prw.sd}}{This slot is a list of unit-specific rw.sd's (the argument of pomp::mif2)}

\item{\code{cooling.type}}{cooling.type}

\item{\code{cooling.fraction.50}}{cooling.fraction.50}

\item{\code{transform}}{transform}

\item{\code{pconv.rec}}{pconv.rec}

\item{\code{pconv.rec.array}}{pconv.rec.array}
}}
\keyword{internal}

