% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelPomp-methods.R
\docType{methods}
\name{names,panelPomp-method}
\alias{names,panelPomp-method}
\title{Get the unit names of a \code{panelPomp} objects.}
\usage{
\S4method{names}{panelPomp}(x)
}
\arguments{
\item{x}{A \code{panelPomp} object.}
}
\description{
Get the unit names of a \code{panelPomp} objects.
}
\details{
S4 method.
}

