% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelGompertz.R
\name{panelGompertz}
\alias{panelGompertz}
\title{Panel Gompertz model}
\usage{
panelGompertz(
  N = 100,
  U = 50,
  params = c(K = 1, r = 0.1, sigma = 0.1, tau = 0.1, X.0 = 1),
  seed = 12345678
)
}
\arguments{
\item{N}{number of observations for each unit.}

\item{U}{number of units.}

\item{params}{parameter vector, assuming all units have the same parameters.}

\item{seed}{passed to the random number generator for simulation.}
}
\value{
\code{panelPomp} object.
}
\description{
Builds a collection of independent realizations from the
Gompertz model.
}
\examples{
panelGompertz()
}
\references{
\breto2020

\king2016
}
\seealso{
Other panelPomp examples: 
\code{\link{contacts}()},
\code{\link{panelRandomWalk}()}
}
\author{
Edward L. Ionides, Carles \Breto
}
\concept{panelPomp examples}
