% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelPomp-class.R
\docType{class}
\name{panelPomp-class}
\alias{panelPomp-class}
\title{An S4 class to represent panel POMP models}
\value{
An S4 \code{panelPomp} class object
}
\description{
An S4 class to represent panel POMP models
}
\section{Slots}{

\describe{
\item{\code{unit.objects}}{A list of \code{pomp} objects}

\item{\code{pParams}}{A \code{list} of length 2 with elements named \code{shared} (a named numeric vector for 
parameters shared among panel units) and \code{specific} (a matrix with row parameter names and column 
unit names). When all parameters are specific, the numeric vector is \code{numeric(0)}. When all parameters 
are shared, the matrix is empty but its number of columns still represents the number of units in the panel 
\code{U}, i.e., \code{array(numeric(0), dim = c(0,U))}.}
}}

\examples{

pompExample(gompertz)

ppo <- panelPomp(
 object = list(
   unit1 = simulate(gompertz, seed = 12345677),
   unit2 = simulate(gompertz, seed = 87654321)
   ),
 shared = c(K = 1.0, r = 0.1, sigma = 0.1),
 specific = matrix(
   rep(c(0.1, 1.0), 2),
   ncol = 2,
   dimnames = list(
     specific.params = c("tau", "X.0"),
     panel.units = paste0("unit", 1:2)
     )
   )
 )
 
 names(ppo@pParams)
 dimnames(ppo@pParams$specific)
 class(ppo)

}
\keyword{internal}
