% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelPomp-construction.R
\docType{methods}
\name{panelPomp,list-method}
\alias{panelPomp,list-method}
\title{Construct \code{panelPomp} objects.}
\usage{
\S4method{panelPomp}{list}(object, shared = numeric(0),
  specific = array(numeric(0), dim = c(0, 0)), params = list(shared =
  shared, specific = specific))
}
\arguments{
\item{object}{a named \code{list} of \code{pomp} objects.}

\item{shared}{a named \code{numeric vector}.}

\item{specific}{a \code{matrix} with parameters on named rows and panel units on named columns.}

\item{params}{optional; a list with (named) 'shared' and 'specific' elements.
more than one column.}
}
\description{
S4 method.
}
\details{
S4 method.
}

