% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelPomp_methods.R
\docType{methods}
\name{panelPomp-methods}
\alias{panelPomp-methods}
\alias{coef,panelPomp-method}
\title{Manipulating \code{panelPomp} objects}
\usage{
\S4method{coef}{panelPomp}(object)
}
\arguments{
\item{object}{An object of class \code{panelPomp} or inheriting class 
\code{panelPomp}.}
}
\description{
Tools for applying particle filtering algorithms to panel data.
}
\section{Methods}{

\describe{
  \item{coef}{Extracts coefficients of \code{panelPomp} objects.}
  \item{pParams}{Extracts coefficients from the \code{pParams} slot of 
  \code{panelPomp} objects.}
  }
}

