% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelPomp-package.R
\docType{package}
\name{panelPomp-package}
\alias{panelPomp-package}
\title{Inference for panel POMPs (Partially Observed Markov Processes)}
\description{
The panelPomp package provides three categories of important functions:
ppomp1, ppomp2 and ppomp3.
}
\section{Constructing a ppomp ogject}{

\code{panelPomp} objects can be constructed calling \code{panelPomp}.


}

