% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelPomp.R
\docType{methods}
\name{panelPomp,panelPomp-method}
\alias{panelPomp,panelPomp-method}
\title{Redefine shared/specific configuration of a panelPomp object.}
\usage{
\S4method{panelPomp}{panelPomp}(object, shared = NULL)
}
\arguments{
\item{object}{panelPomp object.}

\item{shared}{character; names of shared parameters. Those which were not originally shared are copied from the specific parameters for the first panel unit. Defaults to \code{NULL}, which implies all parameters are specific.}
}
\description{
Extract shared and specific parameter values from a panelPomp and use them to create a new panelPomp object with a new configuration of shared and specific parameters.
}
