% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/panelPomp.R
\docType{class}
\name{panelPomp}
\alias{panelPomp}
\alias{panelPomp}
\alias{panelPomp-class}
\alias{panelPomp,list-method}
\alias{panelPomp,panelPomp-method}
\title{\code{panelPomp} generic}
\usage{
panelPomp(object, ...)

\S4method{panelPomp}{list}(object, shared = numeric(0),
  specific = array(numeric(0), dim = c(0, 0)), params = list(shared =
  shared, specific = specific))

\S4method{panelPomp}{panelPomp}(object, shared = NULL)
}
\arguments{
\item{object}{object.}

\item{...}{Additional arguments.}

\item{shared}{either (matching the \code{object} argument above) (i) a named 
\code{numeric vector}; or (ii) a character vector of names for parameters to 
be set to shared. Parameters that were not originally shared are copied from 
the specific parameters from the first panel unit (in the second case, it 
defaults to \code{NULL}, which implies all parameters are specific).}

\item{specific}{a \code{matrix} with parameters on named rows and panel units on named columns.}

\item{params}{optional; a list with (named) 'shared' and 'specific' elements.}

\item{object}{either (i) a named \code{list} of \code{pomp} objects; or (ii) 
an object of class \code{panelPomp} or inheriting class \code{panelPomp}. In  
the second case, the function \code{panelPomp} will manipulate the provided 
object.}
}
\description{
\code{panelPomp} generic

This function constructs a \code{panelPomp} object. It also 
redefine shared/specific configuration of a \code{panelPomp} object.
}
\details{
This is a generic function: methods can be defined for it.
}
\references{
\breto2017

\king2015
}
\seealso{
\pkg{pomp}'s constructor at \link[=pomp]{pomp}

Other panelPomp workhorse functions: \code{\link{mif2}},
  \code{\link{panel_loglik}}, \code{\link{pfilter}}
}
\keyword{internal}
