% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{panelPompExample}
\alias{panelPompExample}
\title{Construct \pkg{panelPomp} examples}
\usage{
panelPompExample(example)
}
\arguments{
\item{example}{a \code{character} object specifying one of the examples that 
come with the \pkg{panelPomp} package. These examples can be listed using 
\code{panelPompExamples()}.}
}
\value{
By default, a \code{panelPomp} object. If \code{envir} is an 
environment, this \code{panelPomp} object is created in that environment 
and named \code{example}.
}
\description{
\code{panelPompExample} constructs \code{panelPomp} objects 
that come with the \pkg{panelPomp} package.
}
\details{
\code{panelPompExample} is related but different than its previous 
(and eliminated as of version 2) \pkg{pomp} counterpart \code{pompExample}.
The \file{examples} directory in the installed package has some 
example files that can be listed using \code{panelPompExamples()}.
}
\author{
Carles Breto and Aaron A. King
}
