% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelRandomWalk.R
\name{panelRandomWalk}
\alias{panelRandomWalk}
\title{Panel random walk model}
\usage{
panelRandomWalk(
  N = 5,
  U = 2,
  params = c(sigmaY = 1, sigmaX = 1, X.0 = 1),
  seed = 3141592
)
}
\arguments{
\item{N}{number of observations for each unit.}

\item{U}{number of units.}

\item{params}{parameter vector, assuming all units have the same parameters.}

\item{seed}{passed to the random number generator for simulation.}
}
\value{
A \code{panelPomp} object.
}
\description{
Builds a collection of independent realizations from a random
walk model.
}
\examples{
panelRandomWalk()
}
\seealso{
Other panelPomp examples: 
\code{\link{contacts}()},
\code{\link{panelGompertz}()}
}
\author{
Edward L. Ionides, Carles \Breto
}
\concept{panelPomp examples}
