% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfilter_methods.R
\name{coef<-,pfilterd.ppomp-method}
\alias{coef<-,pfilterd.ppomp-method}
\alias{shared<-,pfilterd.ppomp-method}
\alias{specific<-,pfilterd.ppomp-method}
\title{Modifying parameters of filtered objects}
\usage{
\S4method{coef}{pfilterd.ppomp}(object, ...) <- value

\S4method{shared}{pfilterd.ppomp}(object) <- value

\S4method{specific}{pfilterd.ppomp}(object) <- value
}
\arguments{
\item{object}{\code{pfilterd.ppomp} object}

\item{...}{additional arguments.}

\item{value}{New parameter value. This function does not allow
users to set this value.}
}
\description{
The setter functions for parameters of \code{pfilterd.ppomp} objects
do not allow users to set parameters of \code{panelPomp} objects that
have been filtered. This is done to avoid the possibility of
having parameter values in an object that do not match other
attributes of a filtered object to be saved together.

The setter functions for parameters of \code{pfilterd.ppomp} objects
do not allow users to set parameters of \code{panelPomp} objects that
have been filtered. This is done to avoid the possibility of
having parameter values in an object that do not match other
attributes of a filtered object to be saved together.

The setter functions for parameters of \code{pfilterd.ppomp} objects
do not allow users to set parameters of \code{panelPomp} objects that
have been filtered. This is done to avoid the possibility of
having parameter values in an object that do not match other
attributes of a filtered object to be saved together.
}
