% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfilterd-ppomp-class.R
\docType{class}
\name{pfilterd.ppomp-class}
\alias{pfilterd.ppomp-class}
\title{An S4 class to represent panel POMP models that have been analyzed using \code{mif2}}
\description{
An S4 class to represent panel POMP models that have been analyzed using \code{mif2}
}
\section{Slots}{

\describe{
\item{\code{Np}}{Np}

\item{\code{ploglik}}{panel loglikelihoods}

\item{\code{ptol}}{panel tolerances}

\item{\code{unit.logliks}}{unit loglikelihoods#'}
}}
\keyword{internal}

