% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-polio.R
\docType{data}
\name{polio_dataset}
\alias{polio_dataset}
\title{Polio data}
\format{An object of class \code{data.frame} with 14112 rows and 5 columns.}
\usage{
polio_dataset
}
\description{
Polio data
}
\keyword{datasets}

