% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{pparams}
\alias{pparams}
\title{\code{pParams} generic.}
\usage{
pparams(object, ...)
}
\arguments{
\item{object}{object.}

\item{...}{Additional arguments.}
}
\description{
\code{pParams} generic function.
}
\details{
This is a generic function: methods can be defined for it.
}
\keyword{internal}
