% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{shared}
\alias{shared}
\title{Extract shared parameters from a panelPomp object}
\usage{
shared(object, ...)
}
\arguments{
\item{object}{an object that contains shared parameters.}

\item{...}{additional arguments.}
}
\value{
vector containing the shared parameters
}
\description{
This function is used to extract the shared parameters from a panelPomp
object.
}
\examples{
prw <- panelRandomWalk()
# extract parameters in list form
shared(prw)
}
\seealso{
\link{panelPomp_methods}
}
\author{
Jesse Wheeler
}
\keyword{internal}
