% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\alias{simulate,panelPomp-method}
\title{Simulations of a panel of partially observed Markov process}
\usage{
\S4method{simulate}{panelPomp}(object, nsim = 1, shared, specific)
}
\arguments{
\item{object}{a \sQuote{panelPomp} object.}

\item{nsim}{The number of simulations to perform. Unlike the pomp
simulate method, all simulations share the same parameters.}

\item{shared}{Named vector of the shared paramters.}

\item{specific}{Matrix of unit-specific parameters, with a column
for each unit.}
}
\value{
A single panelPomp object (if nsim=1) or a list of panelPomp objects
(if nsim>1).
}
\description{
\code{simulate} generates simulations of the state and measurement
processes.
}
\author{
Edward L. Ionides
}
