% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{specific}
\alias{specific}
\title{Extract unit-specific parameters from a panelPomp object}
\usage{
specific(object, ..., format = c("matrix", "vector"))
}
\arguments{
\item{object}{an object that contains unit-specific parameters}

\item{...}{additional arguments.}

\item{format}{character representing how the parameters should be returned}
}
\value{
a matrix or vector containing the unit-specific parameters
}
\description{
This function is used to extract the unit-specific parameters from a
panel pomp object.
}
\examples{
prw <- panelRandomWalk()
# extract parameters in list form
specific(prw)
}
\seealso{
\link{panelPomp_methods}
}
\author{
Jesse Wheeler
}
\keyword{internal}
