% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sQuotes.R
\name{sQuotes}
\alias{sQuotes}
\title{Replace singles quotes (') with \code{sQuote()}s}
\usage{
sQuotes(...)
}
\arguments{
\item{...}{character objects (or other objects to be passed to 
\code{strsplit(paste0(...),"'")[[1]]}).}
}
\description{
Concatenate character vectors and replace singles quotes (') with 
\code{sQuote()}s so that "\code{sQuote('x')}" can be written as just "'x'".
}
\examples{
sQuotes("in 'fn'",": 'object' is a required argument")==
paste0("in ",sQuote('fn'),": ",sQuote('object')," is a required argument")
}
