% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twentycities.R
\docType{data}
\name{twentycities}
\alias{twentycities}
\title{He et al. 2010 twenty UK cities weekly reported measles data}
\format{
\subsection{\code{twentycities}}{

A list of 3 data frames.
}

\subsection{\code{twentycities$measles}:}{

\describe{
\item{unit}{City name}
\item{data}{Date of observation}
\item{cases}{Number of measles cases reported during the week}
}
}

\subsection{\code{twentycities$demog}:}{

\describe{
\item{unit}{City name}
\item{year}{Year that demography was recorded}
\item{pop}{Population}
\item{births}{Births}
}
}

\subsection{\code{twentycities$coord}:}{

\describe{
\item{unit}{City name}
\item{long}{Longitude of city}
\item{lat}{Latitude of city}
}
}
}
\source{
\url{https://kingaa.github.io/pomp/vignettes/twentycities.rda}
}
\usage{
twentycities
}
\description{
He et al. 2010 twenty UK cities weekly reported measles data
}
\keyword{datasets}
