% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{unit_objects}
\alias{unit_objects}
\title{Extract units of a panel model}
\usage{
unit_objects(object, ...)
}
\arguments{
\item{object}{an object for which extraction of panel units is meaningful.}

\item{...}{additional arguments.}
}
\value{
Units extracted from the panel model \code{object}.

\unitobjectsReturn
}
\description{
\code{unit_objects()} is a generic function that extracts a list
of objects corresponding to units of panel objects returned by panel modeling
functions.
}
\details{
This is a generic function: methods can be defined for it.
}
\examples{
prw <- panelRandomWalk()
## access underlying pomp objects
unit_objects(prw)
}
\seealso{
\link{panelPomp_methods}
}
\author{
Carles \Breto
}
\keyword{internal}
