% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfilter_methods.R
\name{unitlogLik,pfilterd.ppomp-method}
\alias{unitlogLik,pfilterd.ppomp-method}
\title{Extract Unit Log-Likelihoods}
\usage{
\S4method{unitlogLik}{pfilterd.ppomp}(object, ...)
}
\arguments{
\item{object}{an object for which log likelihood values for units can be extracted.}

\item{...}{additional arguments.}
}
\value{
When given objects of class \code{pfilterd.ppomp}, \code{unitloglik()} returns a \code{numeric} vector.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
# filter, which generates log likelihoods
pfrw <- pfilter(panelRandomWalk(),Np=10)

# extract log likelihood for each panel unit
unitlogLik(pfrw)

}
