% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelPomp-methods.R
\docType{methods}
\name{unitobjects,panelPomp-method}
\alias{unitobjects,panelPomp-method}
\title{Extract individual \code{pomp} objects from \code{panelPomp} objects.}
\usage{
\S4method{unitobjects}{panelPomp}(object, unitname)
}
\arguments{
\item{object}{A \code{panelPomp} object.}

\item{unitname}{unitname.}
}
\description{
S4 method
}
\details{
S4 method.
}

