% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelPomp-methods.R
\docType{methods}
\name{window,panelPomp-method}
\alias{window,panelPomp-method}
\title{Subset panelPomp objects by changing start time, end time, and number of units.}
\usage{
\S4method{window}{panelPomp}(x, U, start, end)
}
\arguments{
\item{x}{panelPomp object.}

\item{U}{how many units to keep (starting from the first).}

\item{start}{position in original \code{times(pomp)} at which to start the window.}

\item{end}{position in original \code{time(pomp)} at which to end the window.}
}
\description{
Subset panelPomp objects by changing start time, end time, and number of units.
}
