% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelPomp_methods.R
\name{as}
\alias{as}
\title{Coercing \code{panelPomp} objects as \code{list}, \code{pompList} or
\code{data.frame}}
\value{
An object of class matching that specified in the second argument (\code{to=}).
}
\description{
When coercing to a \code{data.frame}, it coerces a
\code{panelPomp} into a \code{data.frame}, assuming units share common
variable names.

When coercing to a \code{list}, it extracts the
\code{unit_objects} slot of \code{panelPomp} objects and attaches
associated parameters.

When coercing to a \code{pompList}, it extracts the
\code{unit_objects} slot of \code{panelPomp} objects and attaches
associated parameters, converting the resulting list to a \code{pompList} to
help the assignment of pomp methods.
}
\seealso{
Other panelPomp methods: 
\code{\link{panelPomp_methods}}
}
\author{
Carles \Breto
}
\concept{panelPomp methods}
