% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelPomp_methods.R
\name{as}
\alias{as}
\title{Coercing \code{panelPomp} objects as a \code{list}}
\description{
Extracts the \code{unit.objects} slot of \code{panelPomp} 
objects.
}
\seealso{
Other panelPomp methods: \code{\link{panelPomp_methods}}
}
