% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelPomp-methods.R
\name{as}
\alias{as}
\title{Extract \code{unit.objects} slot of \code{panelPomp} objects as a \code{list}.}
\description{
Extract \code{unit.objects} slot of \code{panelPomp} objects as a \code{list}.
}
\details{
Extract \code{unit.objects} slot of \code{panelPomp} objects as a \code{list}.
}
