% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelPomp_methods.R
\docType{methods}
\name{coef,panelPomp-method}
\alias{coef,panelPomp-method}
\title{Extract coefficients from the \code{pParams} slot of \code{panelPomp} objects.}
\usage{
\S4method{coef}{panelPomp}(object)
}
\arguments{
\item{object}{A \code{panelPomp} object.}
}
\description{
Extract coefficients from the \code{pParams} slot of \code{panelPomp} objects.
}
\details{
Extract coefficients from the \code{pParams} slot of \code{panelPomp} objects.
}
