% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacts.R
\name{contacts}
\alias{contacts}
\title{Contacts model}
\usage{
contacts(
  params = c(mu_X = 1.75, sigma_X = 2.67, mu_D = 3.81, sigma_D = 4.42, mu_R = 0.04,
    sigma_R = 0, alpha = 0.9)
)
}
\arguments{
\item{params}{parameter vector.}
}
\value{
A \code{panelPomp} object.
}
\description{
A panel model for dynamic variation in sexual contacts, with
data from Vittinghof et al (1999). The model was developed by
Romero-Severson et al (2015) and discussed by \Breto et al (2020).
}
\examples{
\donttest{
contacts()
}
}
\references{
\breto2020

\romeroseverson2015

\vittinghoff1999
}
\seealso{
Other panelPomp examples: 
\code{\link{panelGompertz}()},
\code{\link{panelRandomWalk}()}
}
\author{
Edward L. Ionides
}
\concept{panelPomp examples}
