% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacts.R
\docType{data}
\name{contacts}
\alias{contacts}
\title{Total sexual contacts in four 6-month periods. (Includes only individuals having no missing observations. Authorized for open access. Order has been randomized.)}
\format{A data frame with 882 rows and 5 variables}
\source{
\href{https://github.com/kingaa/sbied/tree/master/contacts}{SBIED Pomp tutorial}
}
\usage{
data(contacts)
}
\description{
Total sexual contacts in four 6-month periods. (Includes only individuals having no missing observations. Authorized for open access. Order has been randomized.)
}
\references{
Vittinghoff, E. et al (1999)  American Journal of Epidemiology 150:306-311
Per-Contact Risk of Human Immunodificiency Virus Tramnsmision between Male Sexual Partners
(\href{https://www.ncbi.nlm.nih.gov/pubmed/10430236}{PubMed})
}
\keyword{datasets}

