% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-exported.R
\name{fromVectorPparams}
\alias{fromVectorPparams}
\title{fromVectorPparams.}
\usage{
fromVectorPparams(vectorPparams)
}
\arguments{
\item{vectorPparams}{A one-row \code{data.frame} with format matching that of the output of \link{toVectorPparams}.}
}
\description{
fromVectorPparams.
}
\details{
Convert back to the format of the \code{pParams} slot \code{panelPomp} objects from a one-row \code{data.frame} (e.g., the row of a record of evaluated likelihoods with the maximum likelihood).
}
