% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_col_row.R
\name{get_dim}
\alias{get_dim}
\alias{get_col}
\alias{get_row}
\title{Get single column or row without dropping names}
\usage{
get_col(matrix, rows, col)

get_row(matrix, row, cols)
}
\arguments{
\item{matrix}{matrix.}

\item{rows}{numeric; rows to subset; like with `[`, this argument can be left empty to designate all rows.}

\item{col}{numeric; single column to subset.}

\item{row}{numeric; single row to subset.}

\item{cols}{numeric; columns to subset; like with `[`, this argument can be left empty to designate all columns.}
}
\description{
Subset matrix dropping dimension but without dropping dimname 
(which is R's default).
}
