% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelGompertz.R
\name{panelGompertz}
\alias{panelGompertz}
\title{Panel Gompertz model}
\usage{
panelGompertz(
  N = 100,
  U = 50,
  params = c(K = 1, r = 0.1, sigma = 0.1, tau = 0.1, X.0 = 1),
  seed = 12345678
)
}
\arguments{
\item{N}{number of observations for each unit.}

\item{U}{number of units.}

\item{params}{parameter vector, assuming all units have the same parameters.}

\item{seed}{passed to the random number generator for simulation.}
}
\description{
Given a collection of points maximizing the likelihood over a range
of fixed values of a focal parameter, this function constructs
a profile likelihood confidence interval accommodating both
Monte Carlo error in the profile and statistical uncertainty present
in the likelihood function.
}
\author{
Edward L. Ionides
}
