% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelPomp-methods-for-construction.R
\docType{methods}
\name{panelPomp,list-method}
\alias{panelPomp,list-method}
\title{Construct \code{panelPomp} objects.}
\usage{
\S4method{panelPomp}{list}(object, shared = numeric(0),
  specific = array(data = numeric(0), dim = c(0, 0)))
}
\arguments{
\item{object}{A \code{list} of \code{pomp} objects.}

\item{shared}{A named \code{numeric vector}}

\item{specific}{A \code{matrix} with parameters on named rows and panel units on named columns
more than one column}
}
\description{
S4 method.
}
\details{
S4 method.
}

