% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelPomp-package.R
\docType{package}
\name{panelPomp-package}
\alias{panelPomp-package}
\title{Inference for panel POMPs (Partially Observed Markov Processes)}
\description{
The panelPomp package helps with inference on panel POMPs.
}
\section{Constructing a ppomp object}{

\code{panelPomp} objects can be constructed by calling \code{panelPomp}.


}

