% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelPomp.R
\docType{class}
\name{panelPomp}
\alias{panelPomp}
\alias{panelPomp-class}
\alias{panelPomp}
\title{Constructing \code{panelPomp} objects}
\usage{
panelPomp(object, shared = new("panelPomp")@pParams$shared,
  specific = new("panelPomp")@pParams$specific, params = list(shared =
  shared, specific = specific))
}
\arguments{
\item{object}{required; either (i) a named \code{list} of \code{pomp} 
objects; or (ii) an object of class \code{panelPomp} or inheriting class 
\code{panelPomp}. In the second case, the function \code{panelPomp} allows 
modifying the provided object.}

\item{shared}{optional; either (matching the \code{object} argument above) 
(i) a named \code{numeric vector}; or (ii) a character vector of names for 
parameters to be set to shared. Parameters that were not originally shared 
are copied from the specific parameters from the first panel unit (in the 
second case, it defaults to \code{NULL}, which implies all parameters are 
specific).}

\item{specific}{optional; a \code{matrix} with parameters on named rows and panel 
units on named columns.}

\item{params}{optional; a list with (named) 'shared' and 'specific' elements.}
}
\description{
This function constructs a \code{panelPomp} object.
}
\references{
\breto2017

\king2015
}
\seealso{
\pkg{pomp}'s constructor at \link[=pomp]{pomp}

Other panelPomp workhorse functions: \code{\link{mif2}},
  \code{\link{panel_loglik}}, \code{\link{pfilter}}
}
