% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{panelPomp}
\alias{panelPomp}
\title{\code{panelPpomp} generic}
\usage{
panelPomp(object, ...)
}
\arguments{
\item{object}{object.}

\item{...}{Additional arguments.}
}
\description{
\code{panelPpomp} generic
}
\details{
This is a generic function: methods can be defined for it.
}
\keyword{internal}
