% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelPomp_methods.R
\docType{methods}
\name{panelPomp_methods}
\alias{panelPomp_methods}
\alias{coef,panelPomp-method}
\alias{coef<-,panelPomp-method}
\alias{length,panelPomp-method}
\alias{names,panelPomp-method}
\alias{pparams,panelPomp-method}
\alias{pParams}
\alias{print,panelPomp-method}
\alias{show,panelPomp-method}
\alias{unitobjects,panelPomp-method}
\alias{window,panelPomp-method}
\alias{[,panelPomp-method}
\alias{[[,panelPomp-method}
\title{Manipulating \code{panelPomp} objects}
\usage{
\S4method{coef}{panelPomp}(object)

\S4method{coef}{panelPomp}(object, ...) <- value

\S4method{length}{panelPomp}(x)

\S4method{names}{panelPomp}(x)

\S4method{pparams}{panelPomp}(object)

pParams(value)

\S4method{print}{panelPomp}(x, ...)

\S4method{show}{panelPomp}(object)

\S4method{unitobjects}{panelPomp}(object)

\S4method{window}{panelPomp}(x, start, end)

\S4method{[}{panelPomp}(x, i)

\S4method{[[}{panelPomp}(x, i)
}
\arguments{
\item{object, x}{An object of class \code{panelPomp} or inheriting class 
\code{panelPomp}.}

\item{...}{....}

\item{value}{value to be assigned.}

\item{start, end}{position in original \code{times(pomp)} at which to start.}

\item{i}{unit index (indices) or name (names).}
}
\description{
Tools for manipulating \code{panelPomp} objects.
}
\section{Methods}{

\describe{
  \item{coef}{Extracts coefficients of \code{panelPomp} objects.}
  \item{coef<-}{Assign coefficients to \code{panelPomp} objects.}
  \item{length}{Count the number of units in \code{panelPomp} objects.}
  \item{names}{Get the unit names of \code{panelPomp} objects.}
  \item{pparams}{Extracts coefficients from \code{panelPomp} objects.}
  \item{[]}{Take a subset of units.}
  \item{[[]]}{Select the pomp object for a single unit.}
  \item{window}{Subset \code{panelPomp} objects by changing start time and
  end time.}
  }
}

\seealso{
Other panelPomp methods: \code{\link{as}}
}
\author{
Carles Breto, Aaron A. King.
}
