% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelPomp_methods.R
\docType{methods}
\name{panelPomp_methods}
\alias{panelPomp_methods}
\alias{coef,panelPomp-method}
\alias{coef<-,panelPomp-method}
\alias{length,panelPomp-method}
\alias{names,panelPomp-method}
\alias{pparams,panelPomp-method}
\alias{pParams}
\alias{unitobjects,panelPomp-method}
\alias{window,panelPomp-method}
\title{Manipulating \code{panelPomp} objects}
\usage{
\S4method{coef}{panelPomp}(object)

\S4method{coef}{panelPomp}(object, ...) <- value

\S4method{length}{panelPomp}(x)

\S4method{names}{panelPomp}(x)

\S4method{pparams}{panelPomp}(object)

pParams(value, object)

\S4method{unitobjects}{panelPomp}(object, unitname)

\S4method{window}{panelPomp}(x, U, start, end)
}
\arguments{
\item{object, x}{An object of class \code{panelPomp} or inheriting class 
\code{panelPomp}.}

\item{...}{....}

\item{value}{value to be assigned.}

\item{unitname}{name of panel unit to be manipulated.}

\item{U}{how many units to keep (starting from the first).
and end the window.}

\item{start, end}{position in original \code{times(pomp)} at which to start}
}
\description{
Tools for manipulating \code{panelPomp} objects.
}
\section{Methods}{

\describe{
  \item{coef}{Extracts coefficients of \code{panelPomp} objects.}
  \item{coef<-}{Assign coefficients to \code{panelPomp} objects.}
  \item{length}{Count the number of units in \code{panelPomp} objects.}
  \item{names}{Get the unit names of \code{panelPomp} objects.}
  \item{pparams}{Extracts coefficients from the \code{pParams} slot of 
  \code{panelPomp} objects.}
  \item{unitobjects}{Extracts \code{pomp} objects from \code{panelPomp} 
  objects.}
  \item{window}{Subset \code{panelPomp} objects by changing start time, 
  end time, and number of units.}
  }
}

\seealso{
Other panelPomp methods: \code{\link{as}}
}
\author{
Carles Breto and Aaron A. King.
}
