% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\alias{toVectorPparams}
\alias{fromVectorPparams}
\alias{toMatrixPparams}
\title{Convert to and from a \code{panelPomp} object \code{pParams} slot format and a one-row \code{data.frame}}
\usage{
toVectorPparams(pParams)

fromVectorPparams(vec_pars)

toMatrixPparams(listPparams)
}
\arguments{
\item{pParams}{A list with the format of the \code{pParams} slot of \code{panelPomp} objects.}

\item{vec_pars}{A one-row \code{data.frame} with format matching that of the
output of \link{toVectorPparams}.}

\item{listPparams}{PanelPomp parameters in list format}
}
\value{
\code{toVectorPparams()} returns an object of class \code{data.frame}.

\code{fromVectorPparams()} returns an object of class \code{list} with the
model parameters in list form.

\code{toMatrixPparams()} returns an object of class \code{matrix} with the
model parameters in matrix form.
}
\description{
These facilitate keeping a record of evaluated log likelihoods.
}
\examples{
prw <- panelRandomWalk()
toVectorPparams(pparams(prw))
fromVectorPparams(toVectorPparams(pparams(prw)))
toMatrixPparams(pparams(prw))
}
\author{
Carles \Breto
}
