% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-exported.R
\name{toVectorPparams}
\alias{toVectorPparams}
\title{toVectorPparams.}
\usage{
toVectorPparams(pParams)
}
\arguments{
\item{pParams}{A list with the format of the \code{pParams} slot of \code{panelPomp} objects.}
}
\description{
toVectorPparams.
}
\details{
Convert from the format of the \code{pParams} slot of \code{panelPomp} objects to a one-row \code{data.frame}. This facilitates keeping a record of evaluated likeilhoods.
}
