% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk_measles.R
\docType{data}
\name{uk_measles}
\alias{uk_measles}
\title{Weekly reported measles data for 355 locations in the UK}
\format{
\subsection{\code{uk_measles}}{

A list of 3 data frames Unit names ending in \code{.RD} are for rural areas; other
unit names are for urban areas. NOTE: not all units have coordinates.
}

\subsection{\code{uk_measles$measles}:}{

\describe{
\item{unit}{City name}
\item{data}{Date of observation}
\item{cases}{Number of measles cases reported during the week}
}
}

\subsection{\code{uk_measles$demog}:}{

\describe{
\item{unit}{City name}
\item{year}{Year that demography was recorded}
\item{pop}{Population}
\item{births}{Births}
}
}

\subsection{\code{uk_measles$coord}:}{

\describe{
\item{unit}{City name}
\item{long}{Longitude of city}
\item{lat}{Latitude of city}
}
}
}
\source{
\url{https://rs.figshare.com/collections/Supplementary_material_from_Structure_space_and_size_competing_drivers_of_variation_in_urban_and_rural_measles_transmission_/5036567/1}
}
\usage{
uk_measles
}
\description{
Weekly reported measles data for 355 locations in the UK
}
\references{
\Korevaar2020
}
\keyword{datasets}
