% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfilterd-ppomp-methods.R
\docType{methods}
\name{unitlogLik,pfilterd.ppomp-method}
\alias{unitlogLik,pfilterd.ppomp-method}
\title{Extract unit loglikelihoods from \code{pfilterd.ppomp} object}
\usage{
\S4method{unitlogLik}{pfilterd.ppomp}(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{Additional arguments}
}
\description{
S4 method
}
\details{
S4 method
}

