% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{unitlogLik}
\alias{unitlogLik}
\title{\code{unitlogLik} generic.}
\usage{
unitlogLik(object, ...)
}
\arguments{
\item{object}{object.}

\item{...}{Additional arguments.}
}
\description{
\code{unitlogLik} generic.
}
\details{
This is a generic function: methods can be defined for it.
}
\keyword{internal}
