% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{unitobjects}
\alias{unitobjects}
\title{\code{unitobjects} generic.}
\usage{
unitobjects(object, ...)
}
\arguments{
\item{object}{object.}

\item{...}{Additional arguments.}
}
\description{
\code{unitobjects} generic.
}
\details{
This is a generic function: methods can be defined for it.
}
\keyword{internal}
