% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wQuotes.R
\name{wQuotes}
\alias{wQuotes}
\title{Interpret shortcuts for \code{sQuote()}s and \code{dQuote()}s in 
character objects}
\usage{
wQuotes(...)
}
\arguments{
\item{...}{character objects (or other objects to be passed to 
\code{strsplit}.}
}
\description{
Concatenate character objects and replace singles quotes with 
\code{sQuote()}s and asterisks with \code{dQuote()}s: \code{sQuote("x")} and
\code{dQuote("x")} can be written as just ''x'' and *x*.
}
\examples{
wQuotes("in ''fn'': *object* is 'a' required argument")
paste0("in ",sQuote("fn"),": ",dQuote("object")," is 'a' required argument")
}
