% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{wQuotes}
\alias{wQuotes}
\title{Interpret shortcuts for \code{sQuote()}s and \code{dQuote()}s in
character objects}
\usage{
wQuotes(...)
}
\arguments{
\item{...}{objects to be passed to \code{strsplit}.}
}
\description{
Concatenate character objects and replace double quotes with
\code{sQuote()}s and asterisks with \code{dQuote()}s: \code{sQuote("x")} and
\code{dQuote("x")} can be written as just ''x'' and *x*.
}
\examples{
wQuotes("in ''fns'': *object* is 'a' required argument")
paste0("in ",sQuote("fns"),": ",dQuote("object")," is 'a' required argument")
}
\author{
Carles Breto
}
\keyword{internal}
